test: extract an integration test visitor as an example plugin
