+++ packages/matchers/package.json
@@ -7,10 +7,13 @@
     "@babel/types": "^7.2.2"
   },
   "devDependencies": {
+    "@babel/core": "^7.2.2",
     "@babel/generator": "^7.2.2",
     "@babel/parser": "^7.2.3",
     "@babel/traverse": "^7.2.3",
+    "@types/babel__core": "^7.0.4",
     "@types/babel__generator": "^7.0.1",
+    "@types/babel__template": "^7.0.1",
     "@types/babel__traverse": "^7.0.4",
     "@types/dedent": "^0.7.0",
     "@types/jest": "^23.3.10",
